last working stuff
