+++ spotify/package.json
@@ -11,6 +11,7 @@
   },
   "dependencies": {
     "@nuxt/eslint": "^1.3.0",
+    "axios": "^1.9.0",
     "bootstrap": "^5.3.5",
     "eslint": "^9.26.0",
     "nuxt": "^3.17.1",
